perf: removed dup functions
