+++ novastack-integrations/observability/watsonx/novastack/observability/watsonx/supporting_classes/utils.py
@@ -1,0 +1,87 @@
+import contextlib
+import io
+import uuid
+from collections.abc import Callable
+from typing import Any
+
+
+def build_payload(
+    records: list[dict],
+    feature_fields: list[str],
+) -> list[dict[str, Any]]:
+    payload_data = []
+    response_fields = ["generated_text", "input_token_count", "generated_token_count"]
+
+    for record in records:
+        request = {"parameters": {"template_variables": {}}}
+        results = {}
+
+        request["parameters"]["template_variables"] = {
+            field: str(record.get(field, "")) for field in feature_fields
+        }
+
+        results = {
+            field: record.get(field) for field in response_fields if record.get(field)
+        }
+
+        pl_record = {
+            "request": request,
+            "response": {"results": [results]},
+            "scoring_id": str(uuid.uuid4()),
+        }
+
+        if "response_time" in record:
+            pl_record["response_time"] = record["response_time"]
+
+        payload_data.append(pl_record)
+
+    return payload_data
+
+
+def suppress_output(fn: Callable, *args: Any, **kwargs: Any) -> Any:
+    """
+    Runs the given function while suppressing all print output (stdout).
+
+    Args:
+        fn: The function to run silently.
+        *args: Positional arguments for the function.
+        **kwargs: Keyword arguments for the function.
+
+    Returns:
+        The return value of the function.
+    """
+    with contextlib.redirect_stdout(io.StringIO()):
+        return fn(*args, **kwargs)
+
+
+def validate_and_filter_dict(
+    original_dict: dict, optional_keys: list, required_keys: list | None = None
+):
+    """
+    Validates that all required keys are present in a dictionary and returns a filtered dictionary
+    containing only the required and specified optional keys with non-None values.
+
+    Args:
+        original_dict (dict): The original dictionary.
+        optional_keys (list): A list of keys to retain.
+        required_keys (list, optional): A list of keys that must be present in the dictionary. Defaults to None.
+    """
+    # Ensure all required keys are in the source dict
+    if required_keys is None:
+        required_keys = []
+
+    missing_keys = [key for key in required_keys if key not in original_dict]
+    if missing_keys:
+        raise KeyError(
+            f"Validation failed: the following required key(s) are missing from the dictionary: {missing_keys}. "
+            "Please provide these keys before proceeding."
+        )
+
+    all_keys_to_keep = set(required_keys + optional_keys)
+
+    # Create a new dictionary with only the key-value pairs where the key is in 'keys' and value is not None
+    return {
+        key: original_dict[key]
+        for key in all_keys_to_keep
+        if key in original_dict and original_dict[key] is not None
+    }
